adding in changes to allow everything to work. still havn't got the resource page to show. i cant find the button
