+++ frontend/mobile-app/package.json
@@ -22,12 +22,14 @@
     "react-native-gifted-chat": "^0.16.1",
     "react-native-ionicons": "^4.6.5",
     "react-native-paper": "^3.10.1",
+    "react-native-phone-call": "^1.0.9",
     "react-native-reanimated": "~1.7.0",
     "react-native-safe-area-context": "0.7.3",
     "react-native-screens": "~2.2.0",
     "react-native-tab-view": "^2.14.4",
     "react-native-vector-icons": "^6.6.0",
-    "react-native-web": "~0.11.7"
+    "react-native-web": "~0.11.7",
+    "react-native-webview": "^10.3.2"
   },
   "devDependencies": {
     "babel-preset-expo": "~8.1.0",
