+++ frontend/mobile-app/screens/ResourceParts/Browser.js
@@ -1,4 +1,4 @@
-import { StatusBar } from 'expo-status-bar';
+//import { StatusBar } from 'expo-status-bar';
 import React from 'react';
 import {
     StyleSheet,
